pyproject.toml
[project]
name = "history-images"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-storage>=3.3.1",
    "pillow>=11.3.0",
    "ray>=2.49.1",
    "tqdm>=4.67.1",
]
